+++ e2e/playwright_test/page/admin/loginAdminPage.po.ts
@@ -1,5 +1,5 @@
 import { Locator, Page, expect } from '@playwright/test';
-import { BasePage } from './basePage.po';
+import { BasePage } from '../basePage.po';
 
 export class LoginAdminPage extends BasePage{
   readonly page: Page;
