+++ e2e/playwright_test/page/admin/overViewAdminPage.po.ts
@@ -1,5 +1,5 @@
 import { Locator, Page } from '@playwright/test';
-import { BasePage } from './basePage.po';
+import { BasePage } from '../basePage.po';
 
 export class OverViewAdminPage extends BasePage{
   readonly page: Page;
